+++ backend/api/messages/reader.go
@@ -1,7 +1,7 @@
 package messages
 
 type ArticleRequest struct {
-	Id uint32 `uri:"id" binding:"required"`
+	Id int32 `uri:"id" binding:"required"`
 }
 
 type ArticleResponse struct {
@@ -11,8 +11,8 @@ type ArticleResponse struct {
 }
 
 type RelatedArticlesRequest struct {
-	ArticleId uint32 `uri:"article_id" binding:"required"`
-	Count     int    `form:"count" binding:"required"`
+	ArticleId int64 `uri:"article_id" binding:"required"`
+	Count     int   `form:"count" binding:"required"`
 }
 
 type RelatedArticlesResponse struct {
